DialogFragment dialog created in onCreateDialog and displayed by show
This is the most preferred way to create a simple dialog. The creation
of the dialog works in the same way as creating a dialog directly
(i.e. with AlertDialog.Builder), but since the dialog is managed by
the DialogFragment, the dialog survives screen rotations and Activity
restarts.

A dialog created in onCreateDialog can be only displayed as a dialog
(i.e. floating window), and not as a Fragment.