DialogFragment dialog created in onCreateView and shown as dialog
The idea of this approach is that with the same implementation of
DialogFragment, the created layout can be displayed both, as a dialog
and as a Fragment (that can be embedded anywhere in the UI).

Whether the layout is displayed as a dialog or as a Fragment is
determined by how the DialogFragment is displayed. For the case of
displaying as a dialog, it is displayed with dialogFrag.show(...).
For the case of displaying as a Fragment, it is treated as a normal
Fragment and added to the UI through a FragmentTransaction.

The drawback of this approach is that the entire dialog UI must be
created manually in XML. It's not possible to create the dialog with
AlertDialog.Builder in onCreateDialog, because onCreateDialog is not
called if the DialogFragment is displayed as a Fragment.

This means that all the details, like padding, text sizes, effects
of positive and negative buttons, must be implemented manually so that
the result resembles a standard AlertDialog.