DialogFragment fragment created in onCreateView and displayed as fragment
The DialogFragment is displayed as a fullscreen fragment, rather than
as a dialog.

The only thing that changes with respect to branch 4_onCreateView_dialog
is how the DialogFragment is displayed in MainActivity (namely by treating
it a s a fragment and adding it in a FragmentTransation, rather than
calling dialogFrag.show(...) to display it as a dialog).